refactor: Make a class for the bot
Cria uma classe para conter a instancia de bot.
Externaliza os handlers para um módulo próprio.
Muda o endpoint para rodar o bot para um script run.py no root.

* Trata de #4
